cuda: widen KQ-mask strides int32->int64 in flash_attn_mask_to_KV_max (#24945)
Manual port of upstream 0eca4d490 to our diverged fattn-common.cuh. Upstream's
single-callsite diff patches the kernel signature + one launch_fattn call site;
our fork has a SECOND call site (the TurboQuant vec-dispatch launch at ~1361)
that upstream does not know about, so all three locations are widened:
- kernel signature (s31/s33 params)
- upstream launch_fattn call site
- fork TQ-dispatch launch site

Without this, 'mask += sequence*s33 + jt*ncols1*s31' truncates in int32 at our
200K-256K ctx + parallel>=2 (n_stream>=2) configs, giving a wrong KV_max scan ->
corrupted attention. Cold once-per-launch kernel, no perf impact, does not touch
the hot TQ vec-dot / V-dequant path.

+++ ggml/src/ggml-cuda/fattn-common.cuh
@@ -604,7 +604,7 @@ static __device__ __forceinline__ void dequantize_V_q8_0(const void * __restrict
 template <int ncols1>
 __launch_bounds__(FATTN_KQ_STRIDE/2, 1)
 static __global__ void flash_attn_mask_to_KV_max(
-        const half2 * __restrict__ mask, int * __restrict__ KV_max, const int ne30, const int s31, const int s33) {
+        const half2 * __restrict__ mask, int * __restrict__ KV_max, const int ne30, const int64_t s31, const int64_t s33) {
     const int ne31     = gridDim.x;
     const int tid      = threadIdx.x;
     const int sequence = blockIdx.y;
@@ -1013,8 +1013,8 @@ void launch_fattn(
     // Only worth the overhead if there is at lease one FATTN_KQ_STRIDE x FATTN_KQ_STRIDE square to be skipped or
     //     multiple sequences of possibly different lengths.
     if (mask && K->ne[1] % FATTN_KQ_STRIDE == 0 && (Q->ne[1] >= 1024 || Q->ne[3] > 1)) {
-        const int s31 = mask->nb[1] / sizeof(half2);
-        const int s33 = mask->nb[3] / sizeof(half2);
+        const int64_t s31 = mask->nb[1] / sizeof(half2);
+        const int64_t s33 = mask->nb[3] / sizeof(half2);
 
         const dim3 blocks_num_KV_max(ntiles_x, Q->ne[3], 1);
         const dim3 block_dim_KV_max(FATTN_KQ_STRIDE/2, 1, 1);
@@ -1358,8 +1358,8 @@ void launch_fattn_paired(
     const int ntiles_dst   = ntiles_x * ntiles_z_gqa * K->ne[2] * Q->ne[3];
 
     if (mask && K->ne[1] % FATTN_KQ_STRIDE == 0 && (Q->ne[1] >= 1024 || Q->ne[3] > 1)) {
-        const int s31 = mask->nb[1] / sizeof(half2);
-        const int s33 = mask->nb[3] / sizeof(half2);
+        const int64_t s31 = mask->nb[1] / sizeof(half2);
+        const int64_t s33 = mask->nb[3] / sizeof(half2);
 
         const dim3 blocks_num_KV_max(ntiles_x, Q->ne[3], 1);
         const dim3 block_dim_KV_max(FATTN_KQ_STRIDE/2, 1, 1);
